Decouple implementations of tuples (staged and runtime) and refactor direct usages of throw QuoteError
