+++ tests/run/tasty-macro-const/quoted_1.scala
@@ -11,10 +11,10 @@ object Macros {
     xTree match {
       case Inlined(_, _, Literal(Constant.Int(n))) =>
         if (n <= 0)
-          throw new QuoteError("Parameter must be natural number")
+          QuoteError("Parameter must be natural number")
         xTree.seal.cast[Int]
       case _ =>
-        throw new QuoteError("Parameter must be a known constant")
+        QuoteError("Parameter must be a known constant")
     }
   }
