+++ tests/run/tuples1b.check
@@ -1,0 +1,4 @@
+0
+(1)
+0
+()
