+++ tests/run/xml-interpolation-1/XmlQuote_1.scala
@@ -18,7 +18,7 @@ object XmlQuote {
     import reflect._
 
     def abort(msg: String): Nothing =
-      throw new QuoteError(msg)
+      QuoteError(msg)
 
     // for debugging purpose
     def pp(tree: Tree): Unit = {
